pyproject.toml
[tool.poetry]
name = "skyblock-stats-bot"
version = "0.1.0"
description = ""
authors = ["mat <github@matdoes.dev>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = {git = "https://github.com/Rapptz/discord.py"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
python-dotenv = "^0.19.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
